Refactor: Remove ViewContext -> Normal Routing
