+++ sales-coach/.env.example
@@ -4,6 +4,8 @@ ANAM_API_KEY=YOUR_ANAM_API_KEY
 # The base URL for the ANAM API
 NEXT_PUBLIC_BASE_URL=https://api.anam.ai
 
+NEXT_PUBLIC_DEFAULT_PERSONA=DEFAULT_PERSONA_ID
+
 # Persona IDs for each language and persona type
 NEXT_PUBLIC_PERSONA_JORDAN=PERSONA_ID
 NEXT_PUBLIC_PERSONA_TAYLOR=PERSONA_ID
