--- sales-coach/src/app/demo/_views/demo/_components/index.ts
@@ -1,2 +1,0 @@
-export * from "./AvatarContainer";
-export * from "./DemoDescription";
