--- sales-coach/src/app/demo/_views/demo/index.ts
@@ -1,1 +1,0 @@
-export * from "./Demo";
