--- sales-coach/src/app/demo/_views/index.ts
@@ -1,3 +1,0 @@
-export * from "./demo";
-export * from "./description";
-export * from "./Scenarios";
