+++ sales-coach/src/app/page.tsx
@@ -7,10 +7,6 @@ import { useRouter } from "next/navigation";
 export default function HomePage() {
   const router = useRouter();
 
-  const handlePlayClick = () => {
-    router.push("/demo");
-  };
-
   return (
     <Box className="appcontainer">
       <Flex justify="start" align="start" className="pt-6">
@@ -21,36 +17,26 @@ export default function HomePage() {
         <Text>
           Practice customer conversations and get personalized feedback
         </Text>
-        <VideoWithPlayButton onPlay={handlePlayClick} />
+        <Box className="relative inline-block">
+          <video
+            className="w-[500px] h-auto"
+            src="leo_gen_1.mp4"
+            poster="leo_gen_1_poster.jpg"
+            autoPlay
+            loop
+            muted
+            playsInline
+          />
+          <IconButton
+            size="4"
+            variant="solid"
+            className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-10"
+            onClick={() => router.push("/scenarios")}
+          >
+            <Play width="30" height="30" />
+          </IconButton>
+        </Box>
       </Flex>
     </Box>
   );
 }
-
-const VideoWithPlayButton = ({
-  onPlay,
-}: {
-  onPlay: (view: string) => void;
-}) => {
-  return (
-    <Box className="relative inline-block">
-      <video
-        className="w-[500px] h-auto"
-        src="leo_gen_1.mp4"
-        poster="leo_gen_1_poster.jpg"
-        autoPlay
-        loop
-        muted
-        playsInline
-      />
-      <IconButton
-        size="4"
-        variant="solid"
-        className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-10"
-        onClick={() => onPlay("scenarios")}
-      >
-        <Play width="30" height="30" />
-      </IconButton>
-    </Box>
-  );
-};
