+++ sales-coach/src/components/DemoControls.tsx
