+++ sales-coach/src/components/GhostButton.tsx
