+++ sales-coach/src/components/Timer.tsx
