+++ sales-coach/src/components/index.ts
@@ -1,1 +1,7 @@
+export * from './AvatarContainer';
+export * from './DemoControls';
+export * from './DemoDescription';
+export * from './GhostButton';
 export * from "./PermissionsModal";
+export * from './Timer';
+export * from './UserVideoContainer';
