+++ sales-coach/src/contexts/AnamContext.tsx
@@ -5,7 +5,7 @@ import constate from "constate";
 import { useRef } from "react";
 import { FetchError, errorHandler, logger, env } from "@/utils";
 
-const PERSONA_ID = env.NEXT_PUBLIC_PERSONA_JORDAN!;
+const PERSONA_ID = env.NEXT_PUBLIC_DEFAULT_PERSONA!;
 const DISABLE_BRAINS = env.NEXT_PUBLIC_DISABLE_BRAINS;
 const DISABLE_FILLER_PHRASES = env.NEXT_PUBLIC_DISABLE_FILLER_PHRASES;
