+++ sales-coach/src/contexts/index.ts
@@ -1,4 +1,3 @@
 export * from "./AnamContext";
 export * from "./VideoAudioPermissionContext";
-export * from "./ViewContext";
 export * from "./SettingsContext";
